pyproject.toml
[project]
name = "tost"
version = "0.2.0"
description = "Token Optimization System Tool — monitor Claude Code session usage"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.9",
    "textual>=0.70",
    "python-dotenv>=1.0",
    "tzdata>=2024.1; sys_platform == 'win32'",
]

[project.scripts]
tost = "tost.cli:main"

[tool.setuptools.packages.find]
include = ["tost*"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
